Make CLI a consumer of formatter
