+++ internal/cmd/fmt.go
@@ -1,16 +1,11 @@
 package cmd
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"path/filepath"
 
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"github.com/stateful/runme/internal/document"
-	"github.com/stateful/runme/internal/document/editor"
-	"github.com/stateful/runme/internal/renderer/cmark"
+	"github.com/stateful/runme/internal/project"
 )
 
 func fmtCmd() *cobra.Command {
@@ -50,59 +45,14 @@ func fmtCmd() *cobra.Command {
 				files = projectFiles
 			}
 
-			for _, relFile := range files {
-				mdFilePath := filepath.Join(proj.Dir(), relFile)
-
-				data, err := readMarkdownFile([]string{mdFilePath})
-				if err != nil {
-					return err
-				}
-
-				var formatted []byte
-
-				if flatten {
-					notebook, err := editor.Deserialize(data)
-					if err != nil {
-						return errors.Wrap(err, "failed to deserialize")
-					}
-
-					if formatJSON {
-						var buf bytes.Buffer
-						enc := json.NewEncoder(&buf)
-						enc.SetIndent("", "  ")
-						if err := enc.Encode(notebook); err != nil {
-							return errors.Wrap(err, "failed to encode to JSON")
-						}
-						formatted = buf.Bytes()
-					} else {
-						formatted, err = editor.Serialize(notebook)
-						if err != nil {
-							return errors.Wrap(err, "failed to serialize")
-						}
-					}
-				} else {
-					doc := document.New(data, cmark.Render)
-					_, astNode, err := doc.Parse()
-					if err != nil {
-						return errors.Wrap(err, "failed to parse source")
-					}
-					formatted, err = cmark.Render(astNode, data)
-					if err != nil {
-						return errors.Wrap(err, "failed to render")
-					}
-				}
-
-				if write {
-					err = writeMarkdownFile([]string{mdFilePath}, formatted)
-				} else {
-					_, _ = fmt.Fprintf(cmd.OutOrStdout(), "===== %s =====\n", relFile)
-					_, _ = cmd.OutOrStdout().Write(formatted)
-					_, _ = fmt.Fprint(cmd.OutOrStdout(), "\n")
-				}
-
-				if err != nil {
-					return err
-				}
+			err = project.Format(files, proj.Dir(), flatten, formatJSON, write, func(file string, formatted []byte) error {
+				_, _ = fmt.Fprintf(cmd.OutOrStdout(), "===== %s =====\n", file)
+				_, _ = cmd.OutOrStdout().Write(formatted)
+				_, _ = fmt.Fprint(cmd.OutOrStdout(), "\n")
+				return nil
+			})
+			if err != nil {
+				return err
 			}
 
 			return nil
